src/requirements.txt
asgiref==3.4.1
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
cryptography==3.4.8
decorator==5.0.9
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.6
django-autoslug==1.9.8
django-braces==1.14.0
django-cors-headers==3.8.0
django-extensions==3.1.3
django-jazzmin==2.4.8
django-oauth-toolkit==1.5.0
django-quill-editor==0.1.22
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
django-taggit==1.5.1
djangorestframework==3.12.4
google-api-core==2.0.1
google-api-python-client==2.19.1
google-auth==2.0.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
ipython==7.27.0
jedi==0.18.0
jwcrypto==1.0
matplotlib-inline==0.1.2
oauthlib==3.1.1
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.20
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
traitlets==5.1.0
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
wrapt==1.12.1
